requirements.txt
python-dotenv
langgraph
langchain
crewai
google-generativeai